test-engineering/contract-tests/client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = "A Python-based test client for the contract tests"
authors = ["Mozilla"]
license = "Mozilla Public License Version 2.0"

[tool.poetry.dependencies]
python = "^3.8.13"
pytest = "^7.2.1"
pydantic = "^1.10.4"
pyyaml = "^6.0"
requests = "^2.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

